more code orginizing
