--- packages/runtime/src/runtime/hydration.ts
@@ -1,3 +1,0 @@
-export class Hydrator {
-    
-}
