--- packages/runtime/src/runtime/runtime.config.ts
@@ -1,4 +1,0 @@
-export class RuntimeConfig {
-    public window:Window = globalThis.window;
-    public document:Document = globalThis.window?.document;
-}
